remove config dep in updateContentBlacklist by extracting api signing logic
