+++ creator-node/src/apiHelpers.js
@@ -1,9 +1,8 @@
 const config = require('./config')
-const Web3 = require('web3')
-const web3 = new Web3()
 
 const { requestNotExcludedFromLogging } = require('./logging')
 const versionInfo = require('../.version.json')
+const { generateTimestampAndSignature } = require('./apiSigning')
 
 module.exports.handleResponse = (func) => {
   return async function (req, res, next) {
@@ -82,45 +81,6 @@ module.exports.successResponse = (obj = {}) => {
     }
   }
 }
-
-/**
- * Generate the timestamp and signature for api signing
- * @param {object} data
- * @param {string} privateKey
- */
-const generateTimestampAndSignature = (data, privateKey) => {
-  const timestamp = new Date().toISOString()
-  const toSignObj = { ...data, timestamp }
-  // JSON stringify automatically removes white space given 1 param
-  const toSignStr = JSON.stringify(sortKeys(toSignObj))
-  const toSignHash = web3.utils.keccak256(toSignStr)
-  const signedResponse = web3.eth.accounts.sign(toSignHash, privateKey)
-
-  return { timestamp, signature: signedResponse.signature }
-}
-module.exports.generateTimestampAndSignature = generateTimestampAndSignature
-
-/**
- * Recover the public wallet address
- * @param {*} data obj with structure {...data, timestamp}
- * @param {*} signature signature generated with signed data
- */
-// eslint-disable-next-line no-unused-vars
-const recoverWallet = (data, signature) => {
-  let structuredData = JSON.stringify(sortKeys(data))
-  const hashedData = web3.utils.keccak256(structuredData)
-  const recoveredWallet = web3.eth.accounts.recover(hashedData, signature)
-
-  return recoveredWallet
-}
-module.exports.recoverWallet = recoverWallet
-
-const sortKeys = x => {
-  if (typeof x !== 'object' || !x) { return x }
-  if (Array.isArray(x)) { return x.map(sortKeys) }
-  return Object.keys(x).sort().reduce((o, k) => ({ ...o, [k]: sortKeys(x[k]) }), {})
-}
-module.exports.sortKeys = sortKeys
 
 const errorResponse = module.exports.errorResponse = (statusCode, message) => {
   return {
